--- a/requirements.txt
+++ b/requirements.txt
@@ -10,3 +10,5 @@ pypng
 Flask-APScheduler
 Pillow
 pynumeral
+arrow
+python-datemath