[`python-datemath`](https://github.com/nickmaccarthy/python-datemath) seems to be a fully compliant, still maintained Python implementation of Kibana's date-math.

Note: it's built on [`arrow`](https://arrow.readthedocs.io/en/latest/), which is Python's version of [`moment.js`](https://momentjs.com/) (i.e., a human-readable date parsing library). 

Also note, it does appear using python-datemath is about 16x slower. See below for some performance metrics:

## Before python-datemath

```python
In [1]: from tms_pixellock import convertKibanaTime                                                                                                     
In [2]: import datetime                                                                                                                                 
In [3]: now = datetime.datetime(2020, 5, 12, 15, 0, 0)                                                                                                  
In [4]: %%timeit 
   ...: convertKibanaTime("now-3m", now) 
                                                                                                        
3.01 µs ± 111 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

## After python-datemath

```python
In [1]: from tms_pixellock import convertKibanaTime                                                                                                     
In [2]: import datetime 
In [3]: now = datetime.datetime(2020, 5, 12, 15, 0, 0) 
In [4]: %%timeit 
   ...: convertKibanaTime("now-3m", now)

48.9 µs ± 2.11 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

This is likely driven by relying on Arrow, which has its own overhead. If this is too much of a performance hit, I can roll the fully-compliant, but not yet fully-tested version I built this morning.