@mrecachinas 2020-05-12
For reference, my implementation (not including the most outer function) is

```
In [8]: from mike_datemath import parse_kibana_time
In [9]: %%timeit 
   ...: parse_kibana_time("-3d", now) 
   ...:  
   ...:                                                                                                                                                 
3.79 µs ± 73.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```