@mrecachinas 2020-05-12
FWIW, my implementation is in the mike-datemath branch.